requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.1
SQLAlchemy==2.0.34
email-validator==2.2.0
WTForms==3.1.2
Flask-WTF==1.2.1

